+++ memberaudit_securegroups/__init__.py
@@ -2,9 +2,4 @@
 Init module
 """
 
-# Standard Library
-from importlib import metadata
-
-__version__ = metadata.version("aa-memberaudit-securegroups")
-
-del metadata
+__version__ = "0.6.0"